targets prune improvement
The targets prune command is a bit hard to use. Its output is
hard to read when you do. This commit uses \n's rather than ,'s to
delimit the target names.

Its also hard to prune things because a target with two tags like
"postmerge,promoted" is hard to build since the logic is to find
an intersection. This change makes the tags you pass have to match
exactly the tags in the target.

Signed-off-by: Andy Doan <andy@foundries.io>