Create a "targets edit" command
This command allows a user to make direct changes to the targets.json
file for a Factory. It tries its best to ensure no garbage is allowed
to be uploaded. However, this command should still be used with
caution.

NOTE: Because this is such a powerful command, it will be hidden
from the "fioctl targets --help" usage command.

Signed-off-by: Andy Doan <andy@foundries.io>