client: Use the UTC aware device listing
The original OTA Lite API was return UTC timestamps but did not indicate
that. In order to prevent breaking older versions of fioctl, the API
was updated to return the non tz-aware timestamps to callers with the
"fioctl" user agent. By changing this value, fioctl will get the new
TZ aware timestamp.

Signed-off-by: Andy Doan <andy@foundries.io>