client: Simplify http client creation
The main reason this is being done is that the current client we
construct won't download big payloads (it stops around 75Mb). While
looking into ways to better construct the client, I realized that
the default client works and has more granular timeout logic in place.

So this simplifies the code and makes it actually work.

Signed-off-by: Andy Doan <andy@foundries.io>