Change pagination logic of device-list
As we grow in the number of devices inside factories, our logic to
show everything in one pass breaks down. The backend now gives us 200
devices at a time. In the case there are more, the "-p <page>" argument
can be added to work through the device. Or, better yet, use filtering
logic.

Signed-off-by: Andy Doan <andy@foundries.io>