common_config: Fix issue with device updates and factory config
This is an interesting bug that's always been present in fioctl. The
idea of this code block was to show what's *changing*. So, if we ran a
command like:

 fioctl devices config updates --apps bar --tags master

and only the apps were changing (tag stayed the same), we'd print:

 Changing apps from [foo] -> [bar]

The problem is that we were also setting the sota toml content in those
conditional blocks. This worked fine for device with:

 # /var/sota/sota.toml
 tag=master

We'd wind up creating an /etc/sota.d/z-50-fioctl.toml with:

 apps=bar

However, if you have done `fioctl config updates --tags=master` and
your /var/sota/sota.toml was something like:

 tags=devel

Running this command would produce a device running on "devel" since
our new z-50-fioctl.toml is missing the tag it *inherits* from the
factory config.

In short - we should always set both fields in the toml.

Signed-off-by: Andy Doan <andy@foundries.io>