Bugfix: fix the IsSliceSetEqual to check for set equality correctly
There was a bug that first and second where deemed as equal if first is a superset of second.
This had no manifestation in the current implementation, as we always replace the offline targets key.
But in the future, it would result into production targets being not signed if e.g. someone adds one more offline targets key.

Signed-off-by: Volodymyr Khoroz <volodymyr.khoroz@foundries.io>