Enhancement: on rotation clean all TUF keys not listed in role spec
The exact behavior in this case is not regulated by TUF specification.
However, aktualizr produces a strange warning on device reboot when it tries to
validate the TUF root which lists the previous TUF root version's key ID in it's keys map.

This change only leaves those keys which are a part of the current TUF root role spec.

Signed-off-by: Volodymyr Khoroz <volodymyr.khoroz@foundries.io>