Feature: use new TUF updates API to get online targets key ID
The existing API only gives the currently active online key ID for the factory.
But we need the key ID which is effective key ID in scope of the TUF updates transaction.
It is not the same if the online targets key was rotated (but the transaction is yet applied).

Signed-off-by: Volodymyr Khoroz <volodymyr.khoroz@foundries.io>