fix(git): Remove sudo
This will remove the need for sudo and walks the user through getting a
writable directory on their PATH so it will all work. Also added
separate Windows logic as the previous one would not work at all in
Windows. For Windows there needs to be one thing set for the user via
Security Policies, which should be added to the docs but I will also
include it here for reference.

Run `secpol.msc` and `Local Policies\User Rights Assignment\Create symbolic links`
should include your current user. Then logoff and login to activate it.

Signed-off-by: Eric Bode <eric.bode@foundries.io>