Feature: ability to view TUF updates in unified diff format
I've tried several packages for this.

The most powerful is the github.com/sergi/go-diff/diffmatchpatch,
which is quite efficient and powerful, and supports the color highlighting.
But is does not support the unified diff, making it look somewhat blurry.

The github.com/karrick/godiff is no longer maintained,
but it is extremely simple and just does exactly what we need.

Signed-off-by: Volodymyr Khoroz <volodymyr.khoroz@foundries.io>