Bump golangci-ling to 1.51.2
I'm helping with LmP upgrade of Golang to 1.20 for fioconfig.
So, my system-wide Golang version was upgraded to version 1.20.
But, for some weird reason, older golangci-lint stopped working after that
regardless of the go1.19 flag in go.mod.
It looks like it uses some of the Golang internals.
The easiest solution is to upgrade golangci-lint to a newer version,
which support Golang 1.17+.

Looking ahead, this is a good addition anyway to stay tuned.

Signed-off-by: Volodymyr Khoroz <volodymyr.khoroz@foundries.io>