cleanup: reuse just generated key when re-signing prod targets
A logic in original code created a new targets signing key in memory, saved it to creds file,
and then parsed it from creds file to sign prod targets.
Instead, we can use the key stored in memory, there is no need to search for it in a file.

This commit is a preparation to allow multiple signing keys per role,
as some of the lower-level APIs used by this method need to be modified.

Signed-off-by: Volodymyr Khoroz <volodymyr.khoroz@foundries.io>