Feature: preverve prod targets signature threshold on all operations
Currently, we set the production targets signature threshold to 2 on all production TUF root modifications.
As we want to allow setting that value, we need to amend that logic, so it preserves a current value.
We still need to set it to 2 when taking the TUF root offline for the first time.

**Note:**
This means, that older Fioctl versions would break the targets signature threshold.
All users of the factory need to update the Fioctl, if they want to set the threshold value.
This needs to be added to the release notes.

Signed-off-by: Volodymyr Khoroz <volodymyr.khoroz@foundries.io>