Fixup: fix the first offline targets key rotation
Mike Sul found a regression in the recent changes related to multi-key support.
When there is no offline targets key yet - the `rotate-offline-keys --role=targets` fails.

A reason is that in a new code we always require an old key to exist inside the tgz file.
But that condition cannot be met when there are no offline targets keys yet.

A workaround is to use a new command `add-offline-key --role=targets`.
But, we need an existing command to behave the same way in this case for backward-compatibility.

Signed-off-by: Volodymyr Khoroz <volodymyr.khoroz@foundries.io>