Use XGo to enabled the CGo based builds
We reached the point when CGo is a must have to integrate HSM support via PKCS11.
But, the standard Golang toolchain is not able to generate static libraries using CGo.
So, this commit switches to a better toolchain provided by XGo.

There are 2 popular XGo flavors, each with its pros and cons:
- https://github.com/techknowlogick/xgo/
- https://github.com/crazy-max/xgo/

The first one has a bigger number of supported platforms, but bigger size (over 2GB) and is less active.
The second is smaller (1.4GB), is more active, has GitHub docker image, and generates better executable names.

Both work great, so I chose the one with smaller size and GitHub image.

This is not yet a full solution, see the follow-up commits.

Signed-off-by: Volodymyr Khoroz <volodymyr.khoroz@foundries.io>