Makefile: use external linker and enable it debugging
With CGo, static binaries can only be used with external linking.
Anyway, Golang is going to use external linking as soon as we import a package with C extensions.
So, enable is explicitly.

Also, add "-v" option to see the exact command called.
That helped me a lot when debugging how go-link and gcc implicitly inject various options.

Signed-off-by: Volodymyr Khoroz <volodymyr.khoroz@foundries.io>