Makefile: a less convoluted command to write files as current user
This one uses a different provider of the XGo script, which allows to extend the docker args.
We still use the same image from `crazy-max/xgo`, but different bootstrap script.
This way is less convoluted, as it allows to use the entrypoint and volume docker arguments to pass the bootstrap script.

Signed-off-by: Volodymyr Khoroz <volodymyr.khoroz@foundries.io>