Test: add unit tests for the factory PKI
These are the first tests for Fioctl, and not something we usually do.
But, I was rather worried about switching from Bash based to Golang based PKI.
Especially, given a fact that we need to also migrate the HSM support which was broken for ages.

I also worry that it is way too easy to break a feature which is used so rarely, albeit being so important.
So, I would prefer these tests to be run as a part of our CI, meaning that this code is being executed regularly.

This first commit adds non-HSM tests.
As you might notice, it helped me to find a couple of bugs in the Golang based PKI.
It also helped verifying that the new Golang based PKI is identical to our legacy Bash based PKI.
Finally, that both implementations work.

Signed-off-by: Volodymyr Khoroz <volodymyr.khoroz@foundries.io>