targets: offline-update: Check if dest dir already has data
This commit prevents the execution of "targets offlile-update" using a
destination directory that already contains update data (i.e.,
"ostree_repo" or "apps" subdirectories). There is no check for the tuf
subdirectory, since the metadata in that subdir is updated by default,
and there is no conflict.
It is possible to override the default behavior by using a new
--allow-multiple-targets option.

Signed-off-by: Andre Detsch <andre.detsch@foundries.io>