Bugfix: El2Go should not overwrite online device CA file
Right now when user runs `el2go config-device-gateway`,
it will try to overwrite the online device CA file on disk and fail.

It is a rather convoluted kind of the regression:
- Before v0.37 on non-Windows: it worked well due to using the Bash based PKI.
- Before v0.37 on Windows: it would silently overwrite the online device CA.
  This is because the buggy Golang based PKI had this bug initially.
- After v0.37: it would fail when trying to overwrite the online device CA file.
  This is because in v0.37 we made two changes:
  * Switched to the Golang based PKI which had a bug.
  * Changed all PKI implementations to save certificate files as read-only.

This change fixes the Golang based (actually used) PKI to not save El2Go cert to disk.

It also adds the El2Go cert generation to the PKI test.
It was failing in an old Golang implementation on file permissions, and passes now.

Signed-off-by: Volodymyr Khoroz <volodymyr.khoroz@foundries.io>