Bugfix: honor the --pki-dir argument in el2go cert generation
The factory PKI expects all certs/keys to be present in the work dir.
However, the `el2go config-device-gateway` does not set the work dir to --pki-dir arg value.
This change fixes that.

Also, make the `--pki-dir` required, as the user is not expected to keep PKI in the current directory.

Signed-off-by: Volodymyr Khoroz <volodymyr.khoroz@foundries.io>