Feature: allow to revoke device CA
Currently, an API allows to simply remove the device CA from a list of CAs.

This change proposes to implement device CA removal via the X.509 certificate revocation.
That way, a user is required to crypto-sign the operation using the factory root CA.
Thus, we both obtain an cryptographic conscent from the user for this operation and protect against several scenarios:
- An accidental device CA removal by passing a wrong file to "keys ca update".
- An intentional device CA removal by bad admin who does not possess the factory root CA.

After this change, the API will be changed to require the "ca-crt" field to contain all device CAs which:
- Are already present in the "ca-crt" list on the backend.
- And are not explicitly listed in the "ca-revoke-crl" list.

Signed-off-by: Volodymyr Khoroz <volodymyr.khoroz@foundries.io>