Feature: add CRL signing key usage to the factory root CA
We need this in order to be able to sign the CRL for the device CA revocation.

For old factory CAs already in the field, a workaround is to artificially set this bit at runtime.
The API needs to be smart to not check this bit of the factory root CA when validating the CRL.

Signed-off-by: Volodymyr Khoroz <volodymyr.khoroz@foundries.io>