Hotfix: clean up key owners when rotating TUF keys
This is an important hot fix.
Currently, the API fails on subsequent key rotations due to denying old key owners.

Fortunately, customers should not see it, as the first rotation succeeds.
But, when the user wants to rotate the key which was rotated using new Fioctl - that second rotation fails.

Signed-off-by: Volodymyr Khoroz <volodymyr.khoroz@foundries.io>