offline-update: Add multi-target bundle support
Add a proper support of multi-target bundles with all required checking
and security. It is achieved by adding an additional signed metadata
file/json that contains an info about a bundle's targets.

- Send `bundle-targets.json` to the `target/meta` handler if exists.
- Receive and store the bundle targets meta if the handler returns them.

The `bundle-targets.json` will be used by the offline updater to
determine targets that can be installed by using a given bundle.

Signed-off-by: Mike Sul <mike.sul@foundries.io>