client: Disable compression in default transport
The default Golang client will advertise to the HTTP server that
it can handle gzip compressed response encoding. This is true,
but it breaks our artifacts.go code for doing download progress.

This is because the default behavior of the transport is to:

 * provide the proper reader to decompress the response
 * update the content-length header to be -1

The -1 breaks our logic. This fix gets us back to how things used
to work before CloudFlare started compressing the responses.

Signed-off-by: Andy Doan <andy@foundries.io>