Cleanup: use net/url to escape special characters in device list
Currently, we onlhy special-cased the URL escaping of a ? for device name pattern.
But, other characters are not escaped, and uuid pattern is not escaped at all.

Most of the time this just works, but may produce unexpected results.
For example `fioctl devices list 'a*&'` will ignore the trailing ampersand,
and behave just like `fioctl device list 'a*'`, matching unexpected devices.

Using the `net/url` to URL escape the query makes filtering more predictable for such edge cases.

Signed-off-by: Volodymyr Khoroz <volodymyr.khoroz@foundries.io>