targets: Handle absence of apps for offline update
Apps might be explicitly turned off or intersection of target apps and
the app shortlist specified for offline update is empty (for specific
tag). In this case there is no a download URI to fetch an app tarball
for a given bundle. This change handles this case properly instead of
yielding an error.

Signed-off-by: Mike Sul <mike.sul@foundries.io>