Feature: a command to start root CA renewal process (w/o HSM)
This adds the first workflow command of the root CA renewal,
which generates the EST compliant CA renewal bundle and uploads it to the server.

This is a bare minimum implementation, further extended with auxiliary features in later commits.
For example, an HSM support is added in the next commit.
That approach allows to decrease the level of complexity while traversing commits.

Signed-off-by: Volodymyr Khoroz <volodymyr.khoroz@foundries.io>