Feature: commands to trigger root CA renewal on devices
This closely resembles the `rotate-cert` commands layout by adding two new commands:
- A `devices config renew-root` for device level config.
- A `config renew-root` for group and factory level configs.

The key difference is the support for factory-wide config change.
That is needed to facilitate the root CA renewal for group-less devices and devices (auto-)registered in the future.

The created config needs a correlation ID which is fetched from the server.
The server generates a new correlation ID upon any changes to the root CA bundle; entire change log is stored for audit.
This is needed so that config updates triggered for the same root CA renewal (but different devices) are counted as one.
That allows to accurately calculate a number of already updated devices, before proceeding to the next root CA renewal step.

I am not sure if the proposed layout is the best one from the user convenience perspective.
The other option was to add single command like `keys ca renewal deploy [-g | -d]`.
That has its pros and cons, so I am open to start a discussion on it.

Signed-off-by: Volodymyr Khoroz <volodymyr.khoroz@foundries.io>