fix(git): throw error when using ~
When using it like `fioctl configure-git --creds-path ~/.config` there
is no problem however if you use it like `fioctl configure-git
--creds-path=~/.config` there is.

The reason lies in the fact that in the first example the shell already
expands the ~ into the absolute path of the homedir, in the second it
will not.

Let's throw an error the same way golang does itself.
https://github.com/golang/go/commit/8b6534b78af791b80f371857a15d76bbc10fd012#diff-7c289d314f17f4e276724c62cfd8dd5e2c7dffad2d900987f8eae83192ad15bdL131

Signed-off-by: Eric Bode <eric.bode@foundries.io>