targets: Add missing fields to TufCustom
The `fioctl targets add` command must retain the arch and image-file
custom fields from the previous target. Otherwise, the
publish-compose-apps operation for new targets will fail.

This commit adds both field to the structure used in the process,
allowing the required fields to be copied to the target being added.

Signed-off-by: Andre Detsch <andre.detsch@foundries.io>