configure-git: Improve logic for setting up helper
Too much of this original work copied directly from the docker
credential helper. It turns out git allows you to specify the full path
to your credential helper. This really simplifies our work because we no
longer need to find a place we can create a symlink with. We simply find
our fioctl executable and use that path.

We also do some extra work to find the config file so that this command
will "just work" with MEDs deployments.

Signed-off-by: Andy Doan <andy@foundries.io>