Upgrade golang from 1.19 to 1.24 (#466)
* Upgrade Golang to 1.24

Our current 1.19 version is no longer supported.
Latest stable is 1.24 - so upgrade to it.

* Upgrade golangci-lint to v2.1.6

Our version is no longer supported and blocks GitHub action upgrades.
This upgrade required 2 more changes:
- A new .golangci.yml format.
- Removed support of --version and --out-format flags.

* Linter: alias non-canonical imports

New linter does not understand imports like go-smth or smth.v2.
They need to be prefixed with an alias, otherwise they produce errors.

* Linter: disable some linter warnings which are not applicable.

New linter introduced a lot of useful checks.
Some of them are too opinionated and not applicable to our use case.

* Cleanup: use RevokedCertificateEntries instead of RevokedCertificates.

This is a cleaned way to work with CRLs since Golang 1.21.
Old way was deprecated in 1.22 producing a linter warning.

* Cleanup: use crypto/ecdh instead of crypto/elliptic to marshal public key bytes.

An elliptic.Marshal has been deprecated since Go 1.21: for ECDH, use the crypto/ecdh package.
This function returns an encoding equivalent to that of PublicKey.Bytes in crypto/ecdh.

* Cleanup: minor linter corrections

Signed-off-by: Volodymyr Khoroz <volodymyr.khoroz@foundries.io>