devices updates: Fix `by-uuuid-flag` warning
We tried adding support for `-by-uuid` to all device subcommands:
  https://github.com/foundriesio/fioctl/commit/89be418dfcded2565a7357c92453fb8840b85ab5

However, the logic in `addUuidFlagToChildren` only adds the flag to
commands that have subcommands. The `updates` command is both its own
command and also has a deprecated `show` subcommand. The flag winds up
only getting added to `updates show` and not `updates`.
```
ERROR: flag accessed but not defined: by-uuid
```

This `updates show` command was deprecated in 2021. By removing that,
the flag will be fixed.

Signed-off-by: Andy Doan <andy@foundries.io>