config: Fix pattern matching for tags
We have different rules for configuring tag names than app names. This
change pulls in the regular expression used by our backend to validate
tag names. This will allow users to configure a device to follow a tag
like `foo.bar` which was already possible, but fioctl would not allow.

Signed-off-by: Andy Doan <andy@foundries.io>